refactor: extract all activity routes into activity_bp Blueprint
Move ~64 activity routes from app.py into activity/routes.py with
a dedicated activity_bp Blueprint. app.py is now a thin 34-line
orchestrator that only creates the Flask app and registers blueprints.
